pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "pythonbotversion-10"
version = "0.1.0"
[tool.poetry.dependencies]
asyncpg = "^0.20.1"
bs4 = "^0.0.1"
config = "^0.5.0"
discord = "^1.0"
django-tango = "^0.10.0"
emoji = "^0.5.4"
flask = "^1.1"
googletrans = "^2.4"
imageio = "^2.8"
lxml = "^4.5"
pkg-info = "^0.1.2"
praw = "^7.0"
psutil = "^5.7"
pyfiglet = "^0.8.0"
pyshorteners = "^1.0"
python = "^3.8"
utils = "^1.0"
wikipedia = "^1.4"
wordcloud = "^1.7"
youtube-dl = "^2020.5"
disputils = "^0.1.3"
paginator = "^0.5.1"
boto = "^2.49"
danksearch = "^0.0.8"
"nekos.py" = "^1.0.3"
ezjson = "^1.0"
python-box = "^4.2.3"
humanize = "^2.5.0"
client = "^0.0.1"
imdb-cli-tool = "^1.0.1"
algorithm = "^1.2.0"
lyricsgenius = "^1.8.6"
validators = "^0.16.0"
python-rapidjson = "^0.9.1"
spotify = "^0.10.2"
covid = "^2.3.0"
pytimeparse = "^1.1.8"
qrcode = "^6.1"
google-images-download = "^2.8.0"
"statcord.py" = "^2.2.0"
udpy = "^2.0.0"
ffmpeg-python = "^0.2.0"
buttons = "^0.1.9"
"discord.py" = "^1.4.1"
speedtest-cli = "^2.1.2"
jishaku = "^1.19.0"
wavelink = "^0.9.4"
dblpy = "^0.3.4"
google-api-python-client = "^1.10.0"
[tool.poetry.dev-dependencies]
